+++ linux/ecosystem/testrail/7.0.2.1016/ad/docker-entrypoint.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-echo "[testrail] Starting testrail service"
+echo "[testrail] Welcome to Testrail 7.0.2.1016 with Active Directory plugin"
+
+echo "[testrail] Starting Testrail service"
 
 ##################################################################################