+++ linux/ecosystem/testrail/7.0.2.1016/main/addons/cassandra-php-driver/download.sh
@@ -1,1 +1,0 @@
-wget -r -k -l 7 -p -E -nc --reject="index.html*"  --user-agent=Mozilla/5.0 https://downloads.datastax.com/cpp-driver/